new_changes

+++ my_project.py
@@ -1,4 +1,5 @@
 import sqlite3
+import datetime
 import time
 conn=sqlite3.connect("fourth.db")
 c=conn.cursor()
@@ -9,11 +10,11 @@ def create_table():               #creates table if not present
     
                         #NAME OF TABLE IS hisaab :-)
 def dynamic_data_entry():         #takes data from user at run time
-        print("enter every thing in lower case for future purpose")
-        dte=input("Enter Today's date: ")
-        itm=input("Enter Name of Item: ")
-        qnty=int(input("Enter quantity: "))
-        prce=int(input("Enter Price per piece of Item: "))
+        print(" \n Enter every thing in lower case for future purpose")
+        dte=datetime.date.today()
+        itm=input(" Enter Name of Item: ")
+        qnty=int(input(" Enter quantity: "))
+        prce=int(input(" Enter Price per piece of Item: "))
         totl_price=qnty*prce
         c.execute("INSERT INTO hisaab(date,item,quantity,price,total_price) VALUES(?,?,?,?,?)",
                   (dte,itm,qnty,prce,totl_price))
@@ -22,7 +23,7 @@ def dynamic_data_entry():         #takes data from user at run time
 def total_kharcha():    #to check total expenses
     c.execute("SELECT SUM(total_price) FROM hisaab")
     sum_all=c.fetchone()[0] #PROBLEM HERE (DID NOT KNOW THE PROPER FUNCTIONING OF THIS)
-    print("\nTotal Kharcha till now is : ",sum_all)
+    print("\n Total Kharcha till now is : ",sum_all)
     conn.commit()
         
         
@@ -31,26 +32,31 @@ def read_from_db():                #funtion to read from Database
     data = c.fetchall()
     #print(data)
     for row in data:
-        print(row)
-    time.sleep(2)
+        date=row[0]
+        item=row[1]
+        quantity=row[2]
+        price=row[3]
+        totalprice=row[4]
+        print("\n%-12s %-6s %-6d %-6d %-6d"%(date,item,quantity,price,totalprice))
+        
     
 def delete_all_db():                #delete funtion to clear whole data base
     c.execute('DELETE FROM hisaab')
     conn.commit()
     
 def delete_some_db():               #delete funtion to clear particular ITEM
-    del_itm=input("\nEnter Name of Item You want to delete from storage: ")
+    del_itm=input("\n Enter Name of Item You want to delete from storage: ")
     c.execute("DELETE FROM hisaab WHERE item=?",(del_itm,)) #MOST IMPORTANT SYNTAX   
     conn.commit()
     
 def entry():                        #funtion to make entry in DB multiple times as user requires    
-    num=int(input("Enter number of Items you want to Enter: ")) 
+    num=int(input(" Enter number of Items you want to Enter: ")) 
     for i in range(num):
         dynamic_data_entry()
 
         
 def main_func():                    #main function to execute other funtions        
-    print("MENU")
+    print("\n\tMENU")
     print("\n 1. Data Entry             (Press 1)")
     print(" 2. Check Existing Entry   (Press 2)")
     print(""" 3. Delete a Item from 
@@ -62,8 +68,7 @@ def main_func():                    #main function to execute other funtions
     print(" 5. Check Total Kharcha    (Press 5)")
     print(" 6. EXIT                   (Press 6)")
     
-    ch=int(input('Choice: '))                   
-    print("\n")
+    ch=int(input(' Choice: '))                   
     if(ch==1):
         create_table()
         entry()
@@ -72,18 +77,27 @@ def main_func():                    #main function to execute other funtions
     elif(ch==3):
         delete_some_db()
     elif(ch==4):
-        delete_all_db()
+        check=input("ALL THE DATA WILL BE ERASED\n DO YOU WISH TO PROCEED(Y/N): ")
+        if(check=="Y" or check=="y"):
+         delete_all_db()
+        elif(check=="N" or check=="n"):
+         main_func()
+        else:
+            print(" Please Enter a Proper Value")
+            time.sleep(2)
+            main_func()            
     elif(ch==5):
         total_kharcha()
     elif(ch==6):
-        print("Thankyou!")    
+        print(" Thankyou!")    
         
     else:
-        print("Wrong Choice Try Again!")
+        print(" Wrong Choice Try Again!")
 
 
 
-print("""Application made by - ANUSHIV SHUKLA
+print("""      Application made by - ANUSHIV SHUKLA
+      
       Languages used - Python 3 and Sqlite 3
       
       Application will Run after 5 secs""")
@@ -95,11 +109,12 @@ main_func()
 time.sleep(2)
 
 while True:                 #to stop termination of program without asking USER
-    print("\nWhat to do Now? Final Exit or Continue")
-    print("\n1. Continue       (Press 1)")
-    print("\n2. Final Exit     (Press 2)")
+    time.sleep(2)
+    print("\n What to do Now? Final Exit or Continue")
+    print("\n 1. Continue       (Press 1)")
+    print("\n 2. Final Exit     (Press 2)")
 
-    ch2=int(input("Choice: "))
+    ch2=int(input(" Choice: "))
 
     if(ch2==1):
         main_func()
@@ -108,7 +123,8 @@ while True:                 #to stop termination of program without asking USER
         conn.close()
         break   
     else:
-        print("Wrong Choice Try Again")
+        print("\n Wrong Choice Try Again")
+        time.sleep(2)
